Add list snapshot interval GUI
This patch adds a new GUI component for listing
snapshot intervals.

It also refactors the snapshot GUI code to split it into separate files
as it was getting a bit unwieldly.
